s3 protocl handler split into independent dir
