+++ plugins.go
@@ -5,5 +5,6 @@ import (
 	_ "github.com/koron/nvgd/filter/indexhtml"
 	_ "github.com/koron/nvgd/filter/tail"
 	_ "github.com/koron/nvgd/filter/texttable"
+	_ "github.com/koron/nvgd/protocol/aws"
 	_ "github.com/koron/nvgd/protocol/db"
 )
