+++ protocol/aws/s3.go
@@ -1,4 +1,4 @@
-package protocol
+package aws
 
 import (
 	"bytes"
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/url"
+	"regexp"
 	"strconv"
 	"time"
 
@@ -16,8 +17,15 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/koron/nvgd/config"
 	"github.com/koron/nvgd/ltsv"
+	"github.com/koron/nvgd/protocol"
 	"github.com/koron/nvgd/resource"
 )
+
+const (
+	S3Token = "s3token"
+)
+
+var rxLastComponent = regexp.MustCompile(`[^/]+/?$`)
 
 var s3config = S3Config{
 	Default: S3BucketConfig{},
@@ -33,8 +41,8 @@ var s3ListHandler = &S3ListHandler{
 }
 
 func init() {
-	MustRegister("s3obj", s3ObjHandler)
-	MustRegister("s3list", s3ListHandler)
+	protocol.MustRegister("s3obj", s3ObjHandler)
+	protocol.MustRegister("s3list", s3ListHandler)
 	config.RegisterProtocol("s3", &s3config)
 }
 
@@ -75,22 +83,15 @@ type S3ListHandler struct {
 func (ph *S3ListHandler) Open(u *url.URL) (*resource.Resource, error) {
 	var (
 		bucket = u.Host
-		key    = u.Path
+		prefix = u.Path
 	)
 	conf := ph.Config.bucketConfig(bucket).awsConfig()
 	sess := session.New(conf)
 	svc := s3.New(sess)
-	if len(key) > 0 {
-		key = key[1:]
+	if len(prefix) > 0 {
+		prefix = prefix[1:]
 	}
-	rc, err := ph.listObjects(svc, bucket, key)
-	if err != nil {
-		return nil, err
-	}
-	return resource.New(rc), nil
-}
-
-func (ph *S3ListHandler) listObjects(svc *s3.S3, bucket, prefix string) (io.ReadCloser, error) {
+	// TODO: use continuation token if available.
 	out, err := svc.ListObjectsV2(&s3.ListObjectsV2Input{
 		Bucket:    aws.String(bucket),
 		Prefix:    aws.String(prefix),
@@ -99,6 +100,17 @@ func (ph *S3ListHandler) listObjects(svc *s3.S3, bucket, prefix string) (io.Read
 	if err != nil {
 		return nil, err
 	}
+	rc, err := ph.writeAsLTSV(out, bucket, prefix)
+	if err != nil {
+		return nil, err
+	}
+	rs := resource.New(rc)
+	rs.Put(protocol.ParsedKeys, []string{S3Token})
+	// TODO: embed continuation token to rs if available.
+	return rs, nil
+}
+
+func (ph *S3ListHandler) writeAsLTSV(out *s3.ListObjectsV2Output, bucket, prefix string) (io.ReadCloser, error) {
 	var (
 		buf = &bytes.Buffer{}
 		w   = ltsv.NewWriter(buf, "name", "type", "size", "modified_at", "link", "download")
